# swagger-specs/trading-swagger.json
{
    "swagger": "2.0",
...
    "securityDefinitions": {
        "firebase": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "oidc": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "info": {
        "title": "Composer Trading API",
        "version": "v1.0",
        "description": "API to live trade against symphonies"
    },
    "paths": {
...
        "/api/v1/deploy/market-hours": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Returns market hours schedule for the upcoming week",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "market_hours": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "is_market_open": {
                                                "type": "boolean"
                                            },
                                            "nyse_market_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "market_open": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "market_close": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "required": [
                                            "is_market_open",
                                            "nyse_market_date"
                                        ],
                                        "additionalProperties": false
                                    }
                                }
                            },
                            "required": [
                                "market_hours"
                            ],
                            "additionalProperties": false
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [],
                "summary": "Get market hours",
                "description": "Returns trading hours for the current day and the next week. Includes regular market hours, extended hours, and holidays.",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
        "/api/v1/deploy/accounts/{account-id}/deploy-symphonies": {
            "get": {
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "symphonies": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "symphony_id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "asset_class": {
                                                "description": "Deprecated. See `asset_classes`.",
                                                "type": "string",
                                                "enum": [
                                                    "EQUITIES",
                                                    "CRYPTO"
                                                ]
                                            },
                                            "asset_classes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "enum": [
                                                        "EQUITIES",
                                                        "CRYPTO"
                                                    ]
                                                }
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "color": {
                                                "type": "string"
                                            },
                                            "community_review_status": {
                                                "type": "string",
                                                "x-nullable": true
                                            },
                                            "last_semantic_update_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "rebalance_frequency": {
                                                "type": "string"
                                            },
                                            "rebalance_corridor_width": {
                                                "type": "number",
                                                "format": "double",
                                                "x-nullable": true
                                            },
                                            "is_shared": {
                                                "type": "boolean"
                                            }
                                        },
                                        "required": [
                                            "symphony_id",
                                            "name",
                                            "asset_class",
                                            "asset_classes",
                                            "description",
                                            "color",
                                            "community_review_status",
                                            "last_semantic_update_at",
                                            "rebalance_frequency",
                                            "rebalance_corridor_width",
                                            "is_shared"
                                        ],
                                        "additionalProperties": false
                                    }
                                }
                            },
                            "required": [
                                "symphonies"
                            ],
                            "additionalProperties": false
                        },
                        "description": ""
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "account-id",
                        "description": "Unique identifier (UUID) of the account",
                        "type": "string",
                        "required": true,
                        "format": "uuid"
                    },
                    {
                        "in": "query",
                        "name": "asset_class",
                        "description": "Deprecated.",
                        "type": "string",
                        "required": false,
                        "enum": [
                            "EQUITIES",
                            "CRYPTO"
                        ],
                        "allowEmptyValue": true
                    }
                ],
                "summary": "Get metadata for symphonies which have new invests pending",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
        "/api/v1/deploy/accounts/{account-id}/deploys": {
            "get": {
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "deploys": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "position_id": {
                                                "type": "string",
                                                "x-nullable": true
                                            },
                                            "symphony_id": {
                                                "type": "string",
                                                "x-nullable": true
                                            },
                                            "deploy_created_by": {
                                                "type": "string",
                                                "x-nullable": true
                                            },
                                            "deploy_id": {
                                                "type": "string"
                                            },
                                            "type": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "cash_change": {
                                                "type": "number",
                                                "format": "double",
                                                "x-nullable": true
                                            },
                                            "requested_cash_change": {
                                                "type": "number",
                                                "format": "double",
                                                "x-nullable": true
                                            },
                                            "pending_buying_power": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "required": [
                                            "position_id",
                                            "symphony_id",
                                            "deploy_created_by",
                                            "deploy_id",
                                            "type",
                                            "status",
                                            "cash_change",
                                            "requested_cash_change",
                                            "pending_buying_power",
                                            "created_at"
                                        ],
                                        "additionalProperties": false
                                    }
                                }
                            },
                            "required": [
                                "deploys"
                            ],
                            "additionalProperties": false
                        },
                        "description": ""
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "status",
                        "description": "",
                        "type": "string",
                        "required": false,
                        "enum": [
                            "QUEUED",
                            "CANCELED",
                            "SUCCEEDED",
                            "FAILED",
                            "REJECTED",
                            "EXPIRED"
                        ]
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "",
                        "type": "integer",
                        "required": false,
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "account-id",
                        "description": "Unique identifier (UUID) of the account",
                        "type": "string",
                        "required": true,
                        "format": "uuid"
                    }
                ],
                "summary": "Get all deploy details for this account",
                "description": "Gets details of all deploys, with the same caveats as getting a single deploy.",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
        "/api/v1/deploy/accounts/{account-id}/deploys/{deploy-id}": {
            "get": {
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "position_id": {
                                    "type": "string",
                                    "x-nullable": true
                                },
                                "symphony_id": {
                                    "type": "string",
                                    "x-nullable": true
                                },
                                "deploy_created_by": {
                                    "type": "string",
                                    "x-nullable": true
                                },
                                "deploy_id": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "cash_change": {
                                    "type": "number",
                                    "format": "double",
                                    "x-nullable": true
                                },
                                "requested_cash_change": {
                                    "type": "number",
                                    "format": "double",
                                    "x-nullable": true
                                },
                                "pending_buying_power": {
                                    "type": "boolean"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "required": [
                                "position_id",
                                "symphony_id",
                                "deploy_created_by",
                                "deploy_id",
                                "type",
                                "status",
                                "cash_change",
                                "requested_cash_change",
                                "pending_buying_power",
                                "created_at"
                            ],
                            "additionalProperties": false
                        },
                        "description": ""
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "account-id",
                        "description": "Unique identifier (UUID) of the account",
                        "type": "string",
                        "required": true,
                        "format": "uuid"
                    },
                    {
                        "in": "path",
                        "name": "deploy-id",
                        "description": "Unique identifier for the deployment operation",
                        "type": "string",
                        "required": true
                    }
                ],
                "summary": "Get a deploy's details",
                "description": "Gets details of a deploy. Pending invests may lack a position ID for new positions, while liquidations and account rebalances won't have a cash change, and account rebalances won't have a position or symphony ID.",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            },
            "delete": {
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "account-id",
                        "description": "Unique identifier (UUID) of the account",
                        "type": "string",
                        "required": true,
                        "format": "uuid"
                    },
                    {
                        "in": "path",
                        "name": "deploy-id",
                        "description": "Unique identifier for the deployment operation",
                        "type": "string",
                        "required": true
                    }
                ],
                "summary": "Cancel a pending deploy",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
        "/api/v1/deploy/accounts/{account-id}/symphonies/{symphony-id}/invest": {
            "post": {
                "responses": {
                    "201": {
                        "description": "Successfully queued investment",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "deploy_id": {
                                    "type": "string"
                                },
                                "deploy_time": {
                                    "description": "The auto rebalance deploy's scheduled time, if creating one was necessary.",
                                    "type": "string",
                                    "x-nullable": true,
                                    "format": "date-time"
                                }
                            },
                            "required": [
                                "deploy_id",
                                "deploy_time"
                            ],
                            "additionalProperties": false
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "consumes": [
                    "application/json",
                    "application/transit+msgpack",
                    "application/transit+json",
                    "application/edn"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "symphony-id",
                        "description": "Unique identifier for the Symphony",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "account-id",
                        "description": "Unique identifier (UUID) of the account",
                        "type": "string",
                        "required": true,
                        "format": "uuid"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "amount": {
                                    "description": "Amount of cash to invest or withdraw (in USD)",
                                    "type": "number",
                                    "format": "double"
                                }
                            },
                            "required": [
                                "amount"
                            ],
                            "additionalProperties": false
                        }
                    }
                ],
                "summary": "Invest cash into a Symphony",
                "description": "Add additional capital to a Symphony position. The investment will be queued and executed during the next rebalancing window.",
                "tags": [
                    "Deploy"
                ],
                "security": [
                    {
                        "firebase": []
                    }
                ]
            }
        },
...
    }
}